requirement.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.6
pycrypto==2.6.1
requests==2.18.4
urllib3==1.22
